refactor: 수익률 계산 역할을 LottoResult 에게 부여
